Extracted external content provider.
