+++ src/Catalog/Icons/AttemptResult.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace NuGet.Services.Metadata.Catalog.Icons
+{
+    public enum AttemptResult
+    {
+        Success,
+        FailCanRetry,
+        FailCannotRetry
+    }
+}
