+++ src/Catalog/NuGet.Services.Metadata.Catalog.csproj
@@ -115,14 +115,18 @@
     <Compile Include="Helpers\PackageUtility.cs" />
     <Compile Include="Helpers\Retry.cs" />
     <Compile Include="ICatalogIndexProcessor.cs" />
+    <Compile Include="Icons\AttemptResult.cs" />
     <Compile Include="Icons\CatalogLeafDataProcessor.cs" />
+    <Compile Include="Icons\ExternalIconContentProvider.cs" />
     <Compile Include="Icons\ExternalIconCopyResult.cs" />
     <Compile Include="Icons\HttpResponseMessageProvider.cs" />
     <Compile Include="Icons\ICatalogLeafDataProcessor.cs" />
     <Compile Include="Icons\IconProcessor.cs" />
     <Compile Include="Icons\IconsCollector.cs" />
+    <Compile Include="Icons\IExternalIconContentProvider.cs" />
     <Compile Include="Icons\IHttpResponseMessageProvider.cs" />
     <Compile Include="Icons\IIconProcessor.cs" />
+    <Compile Include="Icons\TryGetResponseResult.cs" />
     <Compile Include="IPackageCatalogItemCreator.cs" />
     <Compile Include="IPackagesContainerHandler.cs" />
     <Compile Include="IHttpRetryStrategy.cs" />
